pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "agent"
version = "0.1.0"
description = "Agent Learning Project"
readme = "README.md"
authors = [
    {name = "Project Author", email = "14680426@qq.com"}
]
license = {text = "MIT"}
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "langchain>=1.0.0,<2.0.0",
    "langchain-openai>=0.1.0,<1.0.0", 
    "langchain-core>=1.0.0,<2.0.0",
    "langchain-community>=0.0.0,<1.0.0",
    "langgraph>=1.0.0,<1.1.0",
    "openai>=1.0.0,<2.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "langgraph-cli>=0.4.7",
    "zhipuai",
    "langgraph-cli[inmem]",
    "sentence_transformers",
    "chromadb",
    "langchain_chroma",
    "langchain_openai",
    "modelscope",
    "torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118",
    "pypdf"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
]
test = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.20.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"langgraph.templates.agent" = "src/agent"
"agent" = "src/agent"

[tool.setuptools.package-data]
"*" = ["*.json", "*.txt", "*.yaml", "*.yml"]